requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.5.2
SQLAlchemy==2.0.36
psycopg2-binary==2.9.9
python-jose==3.3.0
httpx==0.27.2
python-dotenv==1.0.1
azure-storage-blob==12.23.1
email-validator==2.1.0.post1
python-multipart==0.0.9
